Fix tags containing spaces.
The various places where instances of a Tag were used to output HTML
were changed to use the slug attribute of the Tag object.  For example,
the "lan party" tag was linked to "lan%20party.html" and was rendered to
HTML as "lan party" as an id attribute, which isn't valid.  This uses
the slug which inserts a hyphen instead of preserving the space.