Add basic Jekyll configuration
This was generated using:

    jekyll new --skip-bundle --force .

from within the Docker container launched from the image created by
scripts/docker-build.sh.  Most of the comments were removed as they are
something of a tutorial for beginning to use Jekyll.

At this point the Jekyll site is still not functional and content
migration still needs to happen.  Not every generated piece from the
above command is committed here since there is already a Pelican site in
this repository.