Add default Gemfile for Jekyll
After looking at a few other example repositories it seems like a
standard practice to keep the Gemfile and use it to manage the Gem
dependencies needed for the website project.  The unfortunate
side-effect here is that the Gemfile.lock is also committed.  But it is
not "huge" like some package manager lock files.

This became apparent while trying to move the theme files into the
project from the Gem install directory and finding unmet dependencies
because suddenly nothing required them.

The Gemfile also needs to be passed to the docker build context so that
the image building process can incorporate the dependencies.  The
alternative is to update them every time the container starts which
seems unwise as the dependencies will not change frequently.

With the move to using the Gemfile it seemed appropriate to also then
move to using bundle exec instead of running jekyll directly.