Move Pelican metadata to Jekyll front matter
One unfortunate side-effect here is that tags with spaces had to become
tags with hyphens to preserve tags made up of multiple words.  This is
because Jekyll's YAML front matter would not use commas to delimit tags
but a YAML sequence, and that syntax is much uglier than adding hyphens.