Add news/ path to display posts
In the Pelican site this was news.html, news2.html, and so forth.  That
structure is not possible in Jekyll with GitHub Pages, unfortunately.
The standard Jekyll answer for pagination of a directory of numbered
HTML files is the answer.

The directory is news/ and since it is the most dynamic content it
seemed okay to change with the move to Jekyll.

The index.html file is based on the Jekyll documentation [1] with some
customizations.

[1] https://jekyllrb.com/docs/pagination/