Do not use level 1 headings in posts
    find _posts -name '*.md' -exec sed -ri 's/^## /### /g;s/^# /## /g' {} +

It seems that the minima theme does not style h1 elements inside the
post-content class which makes the base h1 style take effect making an
h2 a larger px size than an h1 which just looks bad.

This seems to make sense semantically, or could at least be justified.
And fixing it seemed like layering on more customization than wanted in
the face of reducing the overall maintenance burden.