Stop using Gemfile.lock
Using the lock file was a best practice recommended many places online
for CI tooling around Ruby projects.  However, it created more work for
this project since it just uses the github-pages Gem and whatever that
needs.  As versions changed and updates were made it created dependency
issues that were resolved by simply removing and rebuilding the image.
That workflow can be simplified by ignoring and cleaning up the lock
file whenever it would be generated.