Add subscription instructions for the mailing list
There were some questions about how to subscribe to the mailing list
without having a Google account to join the group through the Google
Groups web interface.  We are only using the mailing list functionality
of the group so anyone can subscribe by sending an email, but that seems
to be less publicized in the current incarnation of Google Groups.